// apps/trade-pro/src/pages/StoreManagement/StockTransferNotes/types.ts
export type TStockTransferNotes = {
  PrintPreview: boolean;
  IsApproved: boolean;
  ApprovedDate: Date;
  DocDate: Date;
  EntryDate: Date;
  ModifyDate: Date;
  ApprovedUserId: number;
  BranchesId: number;
  FinancialYearId: number;
  CompanyId: number;
  DocumentTypeId: number;
  DestinationLocationId: number;
  DocNo: number;
  EntryUserId: number;
  Id: number;
  ModifyUserId: number;
  OrganizationId: number;
  SourceLocationId: number;
  WsRmRequisitionPoId: number;
  RemarksHeader: string;
  ReqStatus: string;
  WsRmStockTransferNotesDetailsList: TWsRmStockTransferNotesDetailsList[];
  WsRmStockTransferNotesExpensesList: TWsRmStockTransferNotesExpensesList[];
  AttachmentsList: null;
  VoucherHeadInvoices: null;
  InventoryStockEvalautionDetailslist: null;
};
export type TWsRmStockTransferNotesDetailsList = {
  BillWeight: number;
  IssuedAmount: number;
  IssuedQty: number;
  IssuedRate: number;
  StockWeight: number;
  ExpenseAmount: number;
  ItemNetAmount: number;
  Id: number;
  ActionTypeId: number;
  ItemId: number;
  ItemUomId: number;
  WarehouseId: number;
  LineId: number;
  WsRmRequisitionPoDetailId: number;
  WsRmRequisitionPoId: number;
  WsRmStockTransferNotesId: number;
  RemarksDetail: string;
  WareHouseName: string;
  ItemName: string;
  UOMCode: string;
  PackEquivalent: number;
  RequisitionNo: number;
  ApprovedQty: number;
};
export type TWsRmStockTransferNotesExpensesList = {
  ItemId: number;
  ItemName: string;
  Qty: number;
  Rate: number;
  Amount: number;
  Remarks: string;
};